All inventory stuff in org.bukkit moved to org.bukkit.inventory
