+++ src/main/java/org/bukkit/entity/ItemDrop.java
@@ -1,6 +1,6 @@
 package org.bukkit.entity;
 
-import org.bukkit.ItemStack;
+import org.bukkit.inventory.ItemStack;
 
 /**
  * Represents a dropped item.
