+++ src/main/java/org/bukkit/event/block/BlockPlaceEvent.java
@@ -1,7 +1,7 @@
 package org.bukkit.event.block;
 
 import org.bukkit.Block;
-import org.bukkit.ItemStack;
+import org.bukkit.inventory.ItemStack;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
