+++ src/main/java/org/bukkit/inventory/Slot.java
@@ -1,4 +1,4 @@
-package org.bukkit;
+package org.bukkit.inventory;
 
 /**
  * Represents a slot in an inventory
